attempting to fix validation errors
Attribute Errors:

For attributes like crossorigin and defer, remove the attribute values if they are meant to be boolean attributes.
Duplicate ID Errors:

Ensure that all id attributes in the HTML files are unique. Duplicate id values like "tab-1", "tab-2", etc., need to be corrected.
Element Hierarchy Errors:

Correct the hierarchy of elements. Elements like <a> inside <button> are not permitted according to the error messages.